feat(admin): light/dark scheme

+++ apps/admin/src/components/layouts/entry.vue
@@ -1,39 +1,40 @@
 <template>
-  <v-card
-    class="mb-5 px-2"
-    :flat="$vuetify.display.mobile"
-    :rounded="$vuetify.display.mobile ? 0 : undefined"
-  >
-    <v-toolbar class="toolbar-items" color="white">
-      <v-btn color="grey-darken-1" :title="$t(`common.action.back`)" :to="{ name: resource.name }" variant="flat">
-        <v-icon icon="$back" start />{{ $t(`common.action.back`) }}
-      </v-btn>
+  <v-card class="mb-2" flat rounded="xl">
+    <v-toolbar class="toolbar-items" color="surface">
+      <v-btn
+        icon="$back"
+        :title="$t(`common.action.back`)"
+        :to="{ name: resource.name }"
+        variant="text"
+      />
+      <v-breadcrumbs v-if="breadcrumbs.length" class="px-1 py-2" :items="breadcrumbs">
+        <template #divider>
+          <v-icon icon="fas fa-caret-right" />
+        </template>
+      </v-breadcrumbs>
+      <v-spacer />
       <v-btn
         v-if="editsResource"
-        color="secondary"
+        color="primary"
+        rounded="pill"
         :title="$t(`common.action.save`)"
-        variant="flat"
         @click="$emit('save')"
       >
         <v-icon icon="$save" start />{{ $t(`common.action.save`) }}
       </v-btn>
       <slot name="actions" />
-      <v-spacer />
       <confirm-dialog
         v-if="canHandleEntry('delete')"
         color="error"
         icon-left="$delete"
         :label="$t('common.action.delete')"
-        :typed-confirm="['surveys'].includes(resource.name) ? entry.name : undefined"
-        variant="flat"
+        :typed-confirm="['surveys'].includes(resource.name) ? entry?.name : undefined"
         @confirm="remove"
       >
-        {{ $t('common.action.confirm.delete', { name: entry.name ? entry.name : entry.id }) }}
+        {{ $t('common.action.confirm.delete', { name: entry?.name ?? entry?.id }) }}
       </confirm-dialog>
     </v-toolbar>
-  </v-card>
-  <v-card :border="!$vuetify.display.mobile" :flat="$vuetify.display.mobile" :tile="$vuetify.display.mobile">
-    <v-tabs bg-color="primary">
+    <v-tabs color="primary">
       <v-tab
         v-for="tab in tabs"
         :key="tab"
@@ -44,13 +45,14 @@
         {{ tabTitle(tab) }}
       </v-tab>
     </v-tabs>
+    <v-divider />
     <slot />
+    <slot name="addons" />
+    <confirm-leave-dialog
+      :model-value="routeLeave"
+      @update:model-value="$emit('update:routeLeave', $event)"
+    />
   </v-card>
-  <slot name="addons" />
-  <confirm-leave-dialog
-    :model-value="routeLeave"
-    @update:model-value="$emit('update:routeLeave', $event)"
-  />
 </template>
 
 <script lang="ts" setup>
@@ -63,12 +65,13 @@ import { has } from 'lodash-es';
 import { computed, inject } from 'vue';
 import { useRouter } from 'vue-router';
 
+import { useBreadcrumbs } from '@intake24/admin/composables/use-breadcrumbs.ts';
 import resources from '@intake24/admin/router/resources';
 import { useHttp } from '@intake24/admin/services';
 import { useMessages, useResource, useUser } from '@intake24/admin/stores';
 import { ConfirmDialog, useI18n } from '@intake24/ui';
 
-import ConfirmLeaveDialog from './confirm-leave-dialog.vue';
+import { ConfirmLeaveDialog } from '../dialogs';
 
 defineOptions({ name: 'EntryLayout' });
 
@@ -79,7 +82,6 @@ const props = defineProps({
   },
   entry: {
     type: Object as PropType<Dictionary>,
-    required: true,
   },
   routeLeave: {
     type: Object as PropType<RouteLeave>,
@@ -101,6 +103,8 @@ const { can } = useUser();
 const http = useHttp();
 const router = useRouter();
 
+const { breadcrumbs } = useBreadcrumbs();
+
 const resourceDef = computed(
   () => resources.find(item => item.name === resource.name),
 );
@@ -110,7 +114,7 @@ const tabs = computed(() => {
   if (isCreate.value)
     return ['create'];
 
-  if (!resourceDef.value)
+  if (!resourceDef.value || !props.entry)
     return [];
 
   const { securables, ownerId } = props.entry;
@@ -124,7 +128,7 @@ const tabs = computed(() => {
 });
 
 function canHandleEntry(action: string) {
-  if (isCreate.value)
+  if (isCreate.value || !props.entry)
     return false;
 
   const { securables, ownerId } = props.entry;
@@ -137,6 +141,9 @@ function tabTitle(tab: string) {
 };
 
 async function remove() {
+  if (!props.entry)
+    return;
+
   const { id, name } = props.entry;
 
   await http.delete(`${resource.api}/${props.id}`);
